face_blur_cycleGAN/requirements.txt
pandas==1.1.5
torchsummary==1.5.1
pymediainfo==5.0.3
faiss-gpu==1.6.5
tensorboardX==2.1.0
tensorboard==2.4.0
tqdm==4.46.0
torch-optimizer
sklearn
albumentations~=1.1.0
numpy==1.23.1
torch~=1.5.1
Pillow~=7.1.2
torchvision~=0.6.0a0+35d732a
matplotlib~=3.4.3
scipy~=1.7.1
ImageHash~=4.2.1
opencv-python==4.5.5.64